data len 16 bit

+++ include/modbus_rtu_slave.h
@@ -90,10 +90,10 @@ enum mbrs_protocol_error {
 /// MODBUS Protocol errors END
 
 // Read function callback type. data* - place answer there, data_len* - place length there
-typedef enum mbrs_protocol_error (mbrs_read_cb_t)(uint16_t address, uint16_t number_of_registers, uint8_t** data, uint8_t* data_len);
+typedef enum mbrs_protocol_error (mbrs_read_cb_t)(uint16_t address, uint16_t number_of_registers, uint8_t** data, uint16_t* data_len);
 
 // Write function callback type. data* - input data buffer
-typedef enum mbrs_protocol_error (mbrs_write_cb_t)(uint16_t address, uint16_t number_of_registers, uint8_t* data, uint8_t data_len);
+typedef enum mbrs_protocol_error (mbrs_write_cb_t)(uint16_t address, uint16_t number_of_registers, uint8_t* data, uint16_t data_len);
 
 // Diagnostic function callback type
 typedef enum mbrs_protocol_error (mbrs_diagnostic_cb_t)(uint16_t subfunction, uint16_t data, uint16_t* return_data);
